# Improves performance of pileup