--- a/src/cljam/pileup/pileup.clj
+++ b/src/cljam/pileup/pileup.clj
@@ -2,90 +2,48 @@
   "Provides simple pileup functions."
   (:require [com.climate.claypoole :as cp]
             [cljam.common :refer [get-exec-n-threads]]
+            [cljam.bam :as bam]
             [cljam.util.sam-util :as sam-util]
             [cljam.io :as io]
-            [cljam.pileup.common :refer [window-width step center]]
+            [cljam.pileup.common :as common]
             cljam.bam.reader)
   (:import cljam.bam.reader.BAMReader))
 
-(defn- update-pile
-  "Updates the pile vector from the alignment, returning the updated pile
-  vector. pile must be a vector."
-  [aln rname pile positions]
-  (if (= rname (:rname aln))
-    (let [win-beg (first positions)
-          win-end (last positions)
-          left (max (:pos aln) win-beg)
-          right (min (sam-util/get-end aln) win-end)]
-      (loop [i left, pile* pile]
-        (if (<= i right)
-          (recur (inc i) (update-in pile* [(- i win-beg)] inc))
-          pile*)))
-    (vec (repeat (count pile) 0))))
-
 (defn- count-for-positions
   "Piles the alignments up and counts them in the positions, returning it as a
-  vector."
-  [alns rname positions]
-  (loop [[aln & rest] alns
-         pile (vec (repeat (count positions) 0))]
-    (if aln
-      (recur rest (update-pile aln rname pile positions))
-      pile)))
+  seq."
+  [alns beg end]
+  (let [pile (long-array (inc (- end beg)))]
+    (doseq [aln alns]
+      (let [left (max (:pos aln) beg)
+            right (min (sam-util/get-end aln) end)
+            left-index (- left beg)]
+        (dotimes [i (inc (- right left))]
+          (aset-long pile (+ i left-index) (inc (aget pile (+ i left-index)))))))
+    (seq pile)))
 
-(defn rpositions
-  "Returns a lazy seq of nums from start (inclusive) to end (inclusive)."
-  ([start end]
-     (rpositions start end start))
-  ([start end n]
-     (if (>= end n)
-       (cons n (lazy-seq (rpositions start end (inc n)))))))
-
-(defn grouped-rpositions
-  "Returns a lazy seq of grouped positions. Equivalent of
-  (partition-all n (rpositions start end)). This implementation is faster."
-  [start end n]
-  (lazy-seq
-   (letfn [(rpositions* [s e]
-             (loop [i e xs `()]
-               (if (>= i s)
-                 (recur (dec i) (cons i xs))
-                 xs)))]
-     (when (<= start end)
-       (cons (rpositions* start (min (+ start n -1) end))
-             (grouped-rpositions (+ start n) end n))))))
-
-(defn- read-alignments
-  "Reads alignments which have the rname and are included in a range defined by
-  the pos and window size, returning the alignments as a lazy seq. Reading
-  depth is shallow."
-  [rdr rname rlength pos]
-  (let [left (max (- pos window-width) 0)
-        right (min rlength (+ pos window-width))]
-    (io/read-alignments rdr {:chr rname
-                             :start left
-                             :end right
-                             :depth :shallow})))
+(defn- regions
+  [start end step]
+  (->> [(inc end)]
+       (concat (range start (inc end) step))
+       (partition 2 1)
+       (map (fn [[s e]] [s (dec e)]))))
 
 (defn- pileup*
   "Internal pileup function."
-  [rdr rname rlength start end]
+  [rdr rname rlength start end step]
   (let [n-threads (get-exec-n-threads)
+        read-fn (fn [r start end]
+                  (io/read-alignments r {:chr rname :start start :end end :deep :shallow}))
         count-fn (fn [xs]
                    (if (= n-threads 1)
-                     (map (fn [[positions alns]]
-                            (count-for-positions alns rname positions)) xs)
+                     (map (fn [[start end]]
+                            (count-for-positions (read-fn rdr start end) start end)) xs)
                      (cp/pmap (dec n-threads)
-                              (fn [[positions alns]]
-                                (count-for-positions alns rname positions)) xs)))]
-    (->> (grouped-rpositions start end step)
-         (map vec)
-         (map (fn [positions]
-                (let [pos (if (= (count positions) step)
-                            (nth positions center)
-                            (nth positions (quot (count positions) 2)))
-                      alns (doall (read-alignments rdr rname rlength pos))]
-                  [positions alns])))
+                              (fn [[start end]]
+                                (with-open [r (bam/clone-reader rdr)]
+                                  (count-for-positions (read-fn r start end) start end))) xs)))]
+    (->> (regions start end step)
          count-fn
          (apply concat))))
 
@@ -105,12 +63,14 @@
   supplied, piles whole range up."
   ([bam-reader rname]
      (pileup bam-reader rname -1 -1))
-  ([bam-reader rname start end]
+  ([bam-reader rname start end & {:keys [step] :or {step common/step}}]
      (try
        (if-let [r (sam-util/ref-by-name (io/read-refs bam-reader) rname)]
-         (pileup* bam-reader
-                  rname (:len r)
-                  (if (neg? start) 1 start)
-                  (if (neg? end) (:len r) end)))
+         (pileup*
+          bam-reader
+          rname (:len r)
+          (if (neg? start) 1 start)
+          (if (neg? end) (:len r) end)
+          step))
        (catch bgzf4j.BGZFException _
          (throw (RuntimeException. "Invalid file format"))))))