This PR improves performance of pileup. (revised version of #40, this uses feature of #39)

There're two main points.
- use primitive arrays
- make step configurable

Average time of ten profiles piling-up BAM file containing about 65000 alignments in chr1:1-4000000.

[serial]
- before `20918 msecs`
- after `801 msecs`

about 26x faster

[parallel]
- before `8631 msecs`
- after `617 msecs`

about 14x faster